third_party/blink/web_tests/platform/generic/fast/css/getPropertyValue-border-expected.txt

1 2 3 4 5 6 7 8 9

Bug 15823: getPropertyValue for border returns null, should compute the shorthand value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS div1.style.getPropertyValue("border") is "5px solid green"
PASS div2.style.getPropertyValue("border") is ""
PASS div3.style.getPropertyValue("border") is ""
PASS div4.style.getPropertyValue("border") is ""
PASS div5.style.getPropertyValue("border") is ""
PASS div6.style.getPropertyValue("border") is ""
PASS div7.style.getPropertyValue("border") is ""
PASS div8.style.getPropertyValue("border") is ""
PASS div9.style.getPropertyValue("border") is ""
PASS successfullyParsed is true

TEST COMPLETE
